Transform  to generic tests
