+++ src/Test/Test.jl
@@ -1,0 +1,12 @@
+module Test
+
+import MutableArithmetics
+const MA = MutableArithmetics
+
+using Test
+
+include("config.jl")
+
+include("int.jl")
+
+end # module
